Trait lta::async::crowd::CrowdRequests

pub trait CrowdRequests<C: Client + ClientExt> {
    // Required methods
    async fn get_passenger_vol_by<S, D>(
        client: &C,
        vol_type: VolType,
        date: D,
        skip: S
    ) -> LTAResult<Vec<String>>
       where S: Into<Option<u32>>,
             D: Into<Option<Date>>;
    async fn get_crowd_density_rt(
        client: &C,
        train_line: MrtLine
    ) -> LTAResult<Vec<StationCrowdLevel>>;
    async fn get_crowd_density_forecast(
        client: &C,
        train_line: MrtLine
    ) -> LTAResult<CrowdDensityForecast>;
}

All APIs pertaining to transportation crowd

Required Methods

async fn get_passenger_vol_by<S, D>( client: &C, vol_type: VolType, date: D, skip: S ) -> LTAResult<Vec<String>>
where S: Into<Option<u32>>, D: Into<Option<Date>>,

Update freq: By 15th of every month, the passenger volume for previous month data will be generated

Note: Link will expire after 5mins!

async fn get_crowd_density_rt( client: &C, train_line: MrtLine ) -> LTAResult<Vec<StationCrowdLevel>>

Returns real-time platform crowdedness level for the MRT/LRT stations of a particular train network line

Update freq: 10 minutes

async fn get_crowd_density_forecast( client: &C, train_line: MrtLine ) -> LTAResult<CrowdDensityForecast>

Returns forecasted platform crowdedness level for the MRT/LRT stations of a particular train network line at 30 minutes interval

Update freq: 24hours

Object Safety

This trait is not object safe.