Module lta::async::traffic

All API pertaining to traffic related data

Functions

get_bike_parking
Returns bicycle parking locations within a radius. The default radius is set as 0.5km

get_carkpark_avail
Returns no. of available lots for HDB, LTA and URA carpark data. The LTA carpark data consist of major shopping malls and developments within Orchard, Marina, HarbourFront, Jurong Lake District. (Note: list of LTA carpark data available on this API is subset of those listed on One.Motoring and MyTransport Portals)

get_erp_rates
Returns ERP rates of all vehicle types across all timings for each zone.

get_est_travel_time
Returns estimated travel times of expressways (in segments).

get_faulty_traffic_lights
Returns alerts of traffic lights that are currently faulty, or currently undergoing scheduled maintenance.

get_road_details
Returns all planned road openings

get_traffic_images
Returns links to images of live traffic conditions along expressways and Woodlands & Tuas Checkpoints.

get_traffic_incidents
Returns current traffic speeds on expressways and arterial roads, expressed in speed bands.

get_traffic_speed_band
Returns current traffic speeds on expressways and arterial roads, expressed in speed bands.

get_vms_emas
Returns traffic advisories (via variable message services) concerning current traffic conditions that are displayed on EMAS signboards along expressways and arterial roads.