Trait lta::async::train::TrainRequests

pub trait TrainRequests<C: Client> {
#[must_use]    fn get_train_service_alert<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<TrainServiceAlert>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;
}

Required methods

#[must_use]fn get_train_service_alert<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<TrainServiceAlert>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns detailed information on train service unavailability during scheduled operating hours, such as affected line and stations etc.

Update freq: ad-hoc

Loading content...