Struct lta::blocking::client::LTAClient

pub struct LTAClient<T> {
    pub api_key: String,
    pub client: T,
    pub base_url: String,
}

A Client to make requests with The Client holds a connection pool internally, so it is advised that you create one and reuse it There are some instance where you might need to customise your client due to certain limitations.

The Client trait has a general constructor method and you should use the reqwest re-export to build you own customised client from the ground up.

Fields
api_key: Stringclient: Tbase_url: String