Trait lta::blocking::taxi::TaxiRequests

pub trait TaxiRequests<C: Client + ClientExt> {
    // Provided methods
    fn get_taxi_avail(
        client: &C,
        skip: impl Into<Option<u32>>
    ) -> LTAResult<Vec<Coordinates>> { ... }
    fn get_taxi_stands(
        client: &C,
        skip: impl Into<Option<u32>>
    ) -> LTAResult<Vec<TaxiStand>> { ... }
}

All APIs pertaining to taxis

Provided Methods

fn get_taxi_avail( client: &C, skip: impl Into<Option<u32>> ) -> LTAResult<Vec<Coordinates>>

Returns location coordinates of all Taxis that are currently available for hire. Does not include “Hired” or “Busy” Taxis.

Update freq: 1min

fn get_taxi_stands( client: &C, skip: impl Into<Option<u32>> ) -> LTAResult<Vec<TaxiStand>>

Returns detailed information of Taxi stands, such as location and whether is it barrier free.

Update freq: Monthly

Object Safety

This trait is not object safe.