lta

Module get_bus_arrival

@generated by satay. Do not edit by hand.

Structs

GetBusArrivalInput

Returns real-time bus arrival information for services at a queried bus stop. The dataset is updated approximately every 20 seconds.

Enums

GetBusArrivalResponse

Returns real-time bus arrival information for services at a queried bus stop. The dataset is updated approximately every 20 seconds.

Functions

decode_get_bus_arrival_response

encode_get_bus_arrival

Returns real-time bus arrival information for services at a queried bus stop. The dataset is updated approximately every 20 seconds.

get_bus_arrival_parts

Returns real-time bus arrival information for services at a queried bus stop. The dataset is updated approximately every 20 seconds.