Trait lta::Client

pub trait Client: Sized {
    type InternalClient;
    type RB;

    // Required methods
    fn new(
        api_key: impl Into<String>,
        client: Self::InternalClient,
        base_url: impl Into<String>
    ) -> Self;
    fn with_api_key(
        api_key: impl Into<String>,
        base_url: impl Into<String>
    ) -> LTAResult<Self>;
    fn req_builder(&self, url: &str) -> Self::RB;
    fn base_url(&self) -> &str;
}

A Client to make requests with The Client holds a connection pool internally, so it is advised that you create one and reuse it

Required Associated Types

type InternalClient

Any backend Client

type RB

Any type that can build requests

Required Methods

fn new( api_key: impl Into<String>, client: Self::InternalClient, base_url: impl Into<String> ) -> Self

General constructor for Self

fn with_api_key( api_key: impl Into<String>, base_url: impl Into<String> ) -> LTAResult<Self>

This method not assign the api_key in struct if the provided key is empty or whitespaces Instead, assign None

fn req_builder(&self, url: &str) -> Self::RB

Returns Self::RB

fn base_url(&self) -> &str

Returns the base URL that is set by user

Object Safety

This trait is not object safe.