Trait lta::async::traffic::TrafficRequests

pub trait TrafficRequests<C: Client> {
#[must_use]    fn get_erp_rates<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<ErpRate>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_carpark_avail<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<CarPark>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_est_travel_time<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<EstTravelTime>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_faulty_traffic_lights<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<FaultyTrafficLight>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_road_details<'life0, 'async_trait>(
        client: &'life0 C, 
        road_details_type: RoadDetailsType, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<RoadDetails>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_traffic_speed_band<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<TrafficSpeedBand>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_traffic_images<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<TrafficImage>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_traffic_incidents<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<TrafficIncident>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_vms_emas<'life0, 'async_trait>(
        client: &'life0 C, 
        skip: Option<u32>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<VMS>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;

#[must_use]    fn get_bike_parking<'life0, 'async_trait>(
        client: &'life0 C, 
        lat: f64, 
        long: f64, 
        dist: Option<f64>
    ) -> Pin<Box<dyn Future<Output = LTAResult<Vec<BikeParking>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait;
}

Required methods

#[must_use]fn get_erp_rates<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<ErpRate>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns ERP rates of all vehicle types across all timings for each zone.

Update freq: Ad-Hoc

#[must_use]fn get_carpark_avail<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<CarPark>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns no. of available lots for HDB, LTA and URA carpark data. The LTA carpark data consist of major shopping malls and developments within Orchard, Marina, HarbourFront, Jurong Lake District. (Note: list of LTA carpark data available on this API is subset of those listed on One.Motoring and MyTransport Portals)

Update freq: 1 min

#[must_use]fn get_est_travel_time<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<EstTravelTime>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns estimated travel times of expressways (in segments).

Update freq: 5min

#[must_use]fn get_faulty_traffic_lights<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<FaultyTrafficLight>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns alerts of traffic lights that are currently faulty, or currently undergoing scheduled maintenance.

Update freq: 2min or whenever there are updates

#[must_use]fn get_road_details<'life0, 'async_trait>(
client: &'life0 C,
road_details_type: RoadDetailsType,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<RoadDetails>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns all planned road openings or road works depending on the RoadDetailsType supplied

Update freq: 24 hours – whenever there are updates

#[must_use]fn get_traffic_speed_band<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<TrafficSpeedBand>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns current traffic speeds on expressways and arterial roads, expressed in speed bands.

Update freq: 5 minutes

#[must_use]fn get_traffic_images<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<TrafficImage>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns links to images of live traffic conditions along expressways and Woodlands & Tuas Checkpoints.

Update freq: 1 to 5 minutes

#[must_use]fn get_traffic_incidents<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<TrafficIncident>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns current traffic speeds on expressways and arterial roads, expressed in speed bands.

Update freq: 5 minutes

#[must_use]fn get_vms_emas<'life0, 'async_trait>(
client: &'life0 C,
skip: Option<u32>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<VMS>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns traffic advisories (via variable message services) concerning current traffic conditions that are displayed on EMAS signboards along expressways and arterial roads.

Update freq: 2 minutes

#[must_use]fn get_bike_parking<'life0, 'async_trait>(
client: &'life0 C,
lat: f64,
long: f64,
dist: Option<f64>
) -> Pin<Box<dyn Future<Output = LTAResult<Vec<BikeParking>>> + Send + 'async_trait>> where
'life0: 'async_trait,

Returns bicycle parking locations within a radius

Dist is default to 0.5 even if you provide None

Update freq: Monthly

Loading content...