Struct lta::blocking::client::LTAClient

pub struct LTAClient { /* fields omitted */ }

A Client to make requests with The Client holds a connection pool internally, so it is advised that you create one and reuse it There are some instance where you might need to customise your client due to certain limitations.

The Client trait has a general constructor method and you should use the reqwest re-export to build you own customised client from the ground up.

Take a look at the reqwest documentation on how to build your own client

Example

use lta::reqwest::blocking::ClientBuilder;
use lta::blocking::client::LTAClient;
use lta::Client;
use std::time::Duration;

fn my_custom_client() -> LTAClient {
    let client = ClientBuilder::new()
        .no_gzip()
        .connect_timeout(Duration::new(420, 0))
        .build()
        .unwrap();

    LTAClient::new(String::from("API_KEY"), client)
}