Trait lta::Client

pub trait Client: Sized {
    type InternalClient;
    type RB;
    fn new<S: Into<String>>(api_key: S, client: Self::InternalClient) -> Self;

    fn with_api_key<S: Into<String>>(api_key: S) -> LTAResult<Self>;

    fn req_builder(&self, url: &str) -> Self::RB;
}

A Client to make requests with The Client holds a connection pool internally, so it is advised that you create one and reuse it

Associated Types

type InternalClient

Any backend Client

type RB

Any type that can build requests

Loading content...
Required methods

fn new<S: Into<String>>(api_key: S, client: Self::InternalClient) -> Self

General constructor for Self

fn with_api_key<S: Into<String>>(api_key: S) -> LTAResult<Self>

This method not assign the api_key in struct if the provided key is empty or whitespaces Instead, assign None

fn req_builder(&self, url: &str) -> Self::RB

Returns Self::RB

Loading content...